tests/redpanda_installer: use RELEASE_LOWER_BOUND for early stop
a global value is needed because releases are saved to a global cache
file.

enstablish a lower bound and stop when reaching values smaller than
this.

the release v21.10.x is the first with bugfixes shippend after the next
release v21.11, and in the test v21.11 is the oldest release actually used