utils/delta_for: assert to std::logic_error
exchange vassert for equivalent std::logic_error exception, with a
backtrace.

the performance profile should be the same, the exception is not meant
to be recoverable but can be intercepted to add context

fuzz testing found an unchecked integer overflow for cases that should
trigger the assert. this is now an exception in debug mode, to be sure
to catch the issue during testing