cs/segment_meta_cstore: is_applicable on all the columns
experience shows that in case of mismatched hints, the first column that
encodes a boolean fails immediatly. this means that the bad path is a
fail-fast path.

on the other hand, running the check on all the columns (as shows in the
benchmarks) is not too costly. the biggest cost is reaching the correct
frame for all the columns