cs/partition_manfiest: _compacted_away_cloud_bytes
Keep track of amount of bytes removed, due to compaction, since the start of the manifest.

Do not serialize this value since it's only used to observe general
behavior of cloud_storage.

the value is increased when an add(segment_meta) operation cause a
replace operation, and the new segment_meta is smaller than the total
replaced size